Fix bug with negative cx_base on chip
We use a negative cx_base to indicate that an axon should not be
used. In this case, we were setting the synapseLen to be 0.
However, it appeared that the synapses would still do something.

To fix this, we now have a dummy set of zero weights that get used
by any axon with a negative cx_base. This ensures they do nothing.