Set PIP_UPGRADE=False but upgrade pip for docs
The issue is that conda installs certifi, and then pip can't
uninstall it because it doesn't detect it.
Preventing pip from upgrading packages is okay since we
disabled caching in nengo-bones.
But we still need to upgrade pip itself because otherwise
we get "Direct url requirement not allowed for dependencies"
when doing `pip install -e .[docs]` for `abr_control`.
This is not a long-term fix, but once we remove conda and then
re-enable caching we shouldn't have this issue.